Commit. Multiple signup/otptoken/credential

+++ fintech-auth/spring/src/main/java/org/acme/fintech/repository/DeviceRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DeviceRepository extends CrudRepository<Device, Integer> {
+public interface DeviceRepository extends CrudRepository<Device, Long> {
 }
