+++ fintech-auth/spring/src/main/java/org/acme/fintech/repository/SignUpRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SignUpRepository extends CrudRepository<SignUp, Integer> {
+public interface SignUpRepository extends CrudRepository<SignUp, Long> {
 }
